Add files via upload
The purpose of this project is to develop a predictive model that can anticipate customer responses based on their behavior or characteristics. Such a model can be valuable in various industries, particularly in marketing and customer relationship management, by enabling businesses to tailor their strategies and offerings to better meet customer needs and preferences. In the context of iFood, the model could help optimize marketing campaigns, personalize recommendations, and improve customer engagement and satisfaction.